creating an actual admin class
This way we can use in Admin controller and ajax controller.

We are starting to get a few, might need a way to auto inject them.
