+++ App/Controllers/Admin/Home.php
@@ -5,7 +5,7 @@ namespace App\Controllers\Admin;
 class Home extends \Core\AdminController {
     public function index(){
 
-        $this->data['userLevel'] = $this->getUserLevel();
+        $this->data['userLevel'] = $this->admin->getUserLevel();
 
         $this->renderView('Admin/Home');
     }
